#9 [refactor] BaseResponse 생성
