+++ app/src/main/java/com/sopt/now/datasource/AuthService.kt
@@ -1,10 +1,8 @@
 package com.sopt.now.datasource
 
 import com.sopt.now.model.login.RequestLoginDto
-import com.sopt.now.model.login.ResponseLoginDto
 import com.sopt.now.model.signup.RequestSignUpDto
-import com.sopt.now.model.signup.ResponseSignUpDto
-import retrofit2.Call
+import com.sopt.now.utils.BaseResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
@@ -13,10 +11,10 @@ interface AuthService {
     @POST("member/join")
     suspend fun signUp(
         @Body request: RequestSignUpDto,
-    ): Response<ResponseSignUpDto>
+    ): Response<BaseResponse<Unit>>
 
     @POST("member/login")
     suspend fun login(
         @Body request: RequestLoginDto
-    ): Response<ResponseLoginDto>
+    ): Response<BaseResponse<Unit>>
 }
