+++ app/src/main/java/com/sopt/now/datasource/InfoService.kt
@@ -1,6 +1,7 @@
 package com.sopt.now.datasource
 
-import com.sopt.now.model.info.ResponseGetInfoDto
+import com.sopt.now.model.info.UserInfo
+import com.sopt.now.utils.BaseResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Header
@@ -9,5 +10,5 @@ interface InfoService {
     @GET("member/info")
     suspend fun getUserInfo(
         @Header("memberId") memberId: String
-    ): Response<ResponseGetInfoDto>
+    ): Response<BaseResponse<UserInfo>>
 }
