+++ app/src/main/java/com/sopt/now/model/info/UserInfo.kt
@@ -4,18 +4,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseGetInfoDto(
-    @SerialName("code")
-    val code: Int,
-
-    @SerialName("message")
-    val message: String,
-
-    @SerialName("data")
-    val data: UserInfo
-)
-
-@Serializable
 data class UserInfo(
     @SerialName("authenticationId")
     val authenticationId: String,
