--- app/src/main/java/com/sopt/now/model/signup/ResponseSignUpDto.kt
@@ -1,13 +1,0 @@
-package com.sopt.now.model.signup
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseSignUpDto(
-    @SerialName("code")
-    val code: Int,
-
-    @SerialName("message")
-    val message: String,
-)
