+++ app/src/main/java/com/sopt/now/utils/BaseResponse.kt
@@ -1,13 +1,16 @@
-package com.sopt.now.model.login
+package com.sopt.now.utils
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseLoginDto(
+data class BaseResponse<T>(
     @SerialName("code")
     val code: Int,
 
     @SerialName("message")
     val message: String,
+
+    @SerialName("data")
+    val data: T? = null
 )
